Finish rename.
