+++ packages/core/src/core/recordingContentGenerator.ts
@@ -16,7 +16,7 @@ import { appendFileSync } from 'node:fs';
 import type { ContentGenerator } from './contentGenerator.js';
 import type { FakeResponse } from './fakeContentGenerator.js';
 import type { UserTierId } from '../code_assist/types.js';
-import { safeJsonStringify } from '../utils/safeJsonStringify.js';
+import { safeJsonStringify } from '../utils/safe-json-helpers.js';
 import type { LlmRole } from '../telemetry/types.js';
 
 // A ContentGenerator that wraps another content generator and records all the
